pyproject.toml
[project]
name = "ai-models"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.116.1",
    "googlemaps>=4.10.0",
    "joblib>=1.5.2",
    "networkx>=3.5",
    "numpy>=2.3.3",
    "openrouteservice>=2.3.3",
    "pandas>=2.3.2",
    "python-dotenv>=1.1.1",
    "requests>=2.32.5",
    "scikit-learn>=1.7.2",
    "supabase>=2.18.1",
    "uvicorn>=0.35.0",
]
